Generalize socket server accept function
The current implementation of the socket server is built specifically around
running a repl, instead of allowing other possible functions to handle input
from the socket. I've mostly pulled out the repl specific code into it's own
function, and allowed for passing an `accept` function (following the same naming
convention as clojure.core.server) to the `open` function in socketRepl.js.

There's still a call to unhookOutputStreams inside of the `close` function that
I believe isn't necessary since this returns the stdout/stdin for the process to
the original value, but we're calling it as the process is being shut down, so
that's cleaned up anyways.

More importantly, any cleanup for a particular accept function should happen
inside that accept function, as we won't be able to generalize the socket server otherwise.