Removed repl specific fn call in close function
From the previous commit, when the process is exiting we call the `close`
function, which in turn calls a repl.js specific function that is supposed to
return stdin/stdout to it's default value. Since we're already exiting the
process anyways, that's going to be undone be node, and we shouldn't need to
clean this ourselves.

More importantly, any cleanup of resources that is specific to a particular
accept function should happen inside of that accept function, and the socket
server should handle just the sockets.