Moved Socket REPL startup into cljs init function
If we want to resolve ClojureScript functions we'll need to have access to the
ClojureScriptContext, which means the Socket REPL initialize has to happen after
the ClojureScript engine is started.

Since the startClojureScriptEngine starts the engine as well as running whatever
scripts, main functions, etc. that the user has started, it made sense to put
the Socket REPL code there as well.