Worked around odd object conversion issue
When building the dev scripts, for some reason passing a JSON object
from the JS code to the CLJS code means we can't convert to cljs objects
with `js->cljs`. This apparently worked if you build the release
executable, but that leaves dev hard to develop/test/debug. So I've
introduced a workaround.