Converted startCLJS to async function
We're using promises to ensure errors in the socket repl don't blow up
the whole repl, so to ensure the server is started before printing it's
details we're using async/await.

Not that I think about it, this should probably go into the socketREPL
namespace, which would change the far reaching nature of these changes.
Hmmm...