Add callback to lumo.build.api/build (#265)
This change was made necessary by the fact that the bootstrap compiler is
asynchronous by default. Therefore, it does not make sense to call
lumo.build.api/build without a callback.

The patch does not break old code, so it still contains the arity without cb,
in which case it hooks a default callback that just returns the input.