Tweak suffix regex in completions
A couple of more tweaks to the suffix regex were necessary.
The first will make so that we consider a : (keyword completion) as suffix and
therefore filter by colon the completion candidates.
The second allows suffixes to have numbers in it - according to Clojure's
specification only the first char of a symbol cannot be a number.