Build lumo against latest node lts (#384)
* Add Travis configuration for two node builds

This patch adds the necessary configuration to build lumo against the latest
node lts, 8.11.1. Against any version for that matter.
The executable name will be different when built in CI: lumo-$nodeVersion.

* Port to CircleCI v2 and build against different node versions

* Add AppVeyor configuration for two node builds

* Make sure port number is correct before creating the socket

This is necessary because jest snapshot testing is not yet ready for testing
multiple node versions and the error message returned by the node net module
changes depending on the version. Therefore, we throw the error ourselves so
that we are in control of the output.